[FIX] Body.js 토큰 헤더 추가
