+++ front/src/views/body/Body_fail.js
@@ -4,8 +4,6 @@ import Menu from 'components/body/menu'
 import Exercise from 'components/body/exercise'
 import Axios from 'axios'
 import Summry from 'components/body/summary'
-import getHeaderCookie from 'utils/hooks/getHeaderCookie'
-import {parseJwt, getMemInfoFromToken} from 'utils/hooks/parseToken'
 
 const BodyMenu = () => {
 
@@ -30,12 +28,7 @@ const BodyMenu = () => {
 
     // Mneu -------------------------------------------------------------------------------------------------
     const [menuData, setMenuData] = useState([]);
-    //Header Cookie
-    const access_token = getHeaderCookie();
-    let parse_token = parseJwt(access_token);
-    let  { memId, memSeq } = getMemInfoFromToken(parse_token);
-    console.log("memId:::::::::::::::::::::::::::::", memId);
-    console.log("memSeq:::::::::::::::::::::::::::::", memSeq);
+
     //식사 유형에 대한 총 칼로리
     const [breakfastTotalCalories, setBreakfastTotalCalories] = useState(0);
     const [lunchTotalCalories, setLunchTotalCalories] = useState(0);
@@ -43,21 +36,33 @@ const BodyMenu = () => {
     const [snackTotalCalories, setSnackTotalCalories] = useState(0);
     const [dailyTotalCalories, setDailyTotalCalories] = useState(0);
     const [menuRefresh, setMenuRefresh] = useState(false);
+    //유재건 쿠키 보내기
+    //유재건
+    const [accessToken, setAccessToken] = useState('');
+    
 
     // 식단 리스트 불러오기
     const callMenu = () => {
-        if (access_token === "") return;
+        const cookies = document.cookie.split(';');
+        let access_token = "";
+        for (const cookie of cookies) {
+            const [name, value] = cookie.trim().split('=');
+            if (name == 'access_token') {
+                access_token = value;
+                break;
+            }
+        }
         console.log(access_token);
-        console.log("memId:::::::::::::::::::::::::::::", memId);
-        console.log("memSeq:::::::::::::::::::::::::::::", memSeq);
+        if(access_token==="") return;
         Axios.get(process.env.REACT_APP_HOST_URL + '/api/menu/list', {
             params: {
                 menu_date: selectDate,
-            }, withCredentials: false, headers: {
-                Authorization: `Bearer ${access_token}`
+            },
+            withCredentials: true,
+            headers: {
+                Authorization: `Bearer ${access_token}`,
             }
         })
-
             .then((res) => {
                 const menuList = res.data.data.menuList;
                 let breakfastCalories = 0;
@@ -106,13 +111,13 @@ const BodyMenu = () => {
         Axios.post(process.env.REACT_APP_HOST_URL + '/api/menu/delete', {
             menuSeq: menu_seq
         })
-            .then((res) => {
-                //alert("삭제 메뉴" + menu_seq);
-                setMenuRefresh(!menuRefresh);
-            })
-            .catch((error) => {
-                console.error(error);
-            });
+        .then((res) => {
+            //alert("삭제 메뉴" + menu_seq);
+            setMenuRefresh(!menuRefresh);
+        })
+        .catch((error) => {
+            console.error(error);
+        });
     }
 
     // Exercise ----------------------------------------------------------------------------------------------
@@ -121,7 +126,7 @@ const BodyMenu = () => {
     const [exerciseRefresh, setExerciseRefresh] = useState(false);
 
     const callExercise = () => {
-        Axios.get(process.env.REACT_APP_HOST_URL + '/api/exercise/list', {
+        Axios.get( process.env.REACT_APP_HOST_URL + '/api/exercise/list', {
             params: {
                 mem_seq: 1,
                 el_date: selectDate,
@@ -136,7 +141,7 @@ const BodyMenu = () => {
                 res.data.data.exerciseList.forEach((item) => {
                     total += item.el_total;
                 });
-
+        
                 setTotalExerciseCalories(total.toFixed(2));
 
             })
@@ -148,7 +153,7 @@ const BodyMenu = () => {
 
     //운동 삭제
     const exerciseDelete = (el_seq) => {
-        Axios.post(process.env.REACT_APP_HOST_URL + '/api/exercise/delete', {
+        Axios.post(process.env.REACT_APP_HOST_URL +'/api/exercise/delete', {
             elSeq: el_seq,
             memSeq: 1
         })
@@ -160,7 +165,7 @@ const BodyMenu = () => {
             });
     }
 
-
+ 
 
     //summary ----------------------------------------------------------------------------------------------------
     const [exTotal, setExTotal] = useState("");             // 운동 총 칼로리
@@ -169,7 +174,7 @@ const BodyMenu = () => {
 
     // 데이터 불러오기
     const callSummary = () => {
-        Axios.get(process.env.REACT_APP_HOST_URL + '/api/summary/list', {
+        Axios.get( process.env.REACT_APP_HOST_URL + '/api/summary/list', {
             params: {
                 mem_seq: 1,
                 menu_date: selectDate,
